+++ src/lunch_menu/db.py
@@ -18,3 +18,21 @@ DB_CONFIG = {
 
 def get_connection():
     return psycopg.connect(**DB_CONFIG)
+
+def insert_menu(menu_name, member_id, dt):
+    try:
+        conn = get_connection()
+        cursor = conn.cursor()
+        cursor.execute(
+                  "INSERT INTO lunch_menu (menu_name, member_id, dt) VALUES (%s,%s,%s);",
+                  (menu_name, member_id, dt)
+                  )
+        conn.commit()
+        cursor.close()
+        conn.close()
+        return True
+    except Exception as e:
+           print(f"Exception:{e}:")
+           return False
+
+
